[EDT-1094] Update internals to move math side to engine
