+++ packages/sdk/package.json
@@ -8,7 +8,6 @@
     },
     "license": "MIT",
     "dependencies": {
-        "mathjs": "^10.0.2",
         "penpal": "6.1.0"
     },
     "devDependencies": {
